requirements.txt
fastapi
pydantic
uvicorn
tuneapi
python-dotenv
jinja2